Make buildjs.sh script build a single minified JS file.
The file name includes a SHA hash for better cacheability.

The process is rather manual at the moment. The coffee.sh script
needs to be run manually after changing the source files,
the `JS_SHA` variable in `views.py` needs to be manually updated
every time the contents of the minified JS file change,
and the list of JS source files in `buildjs.sh` needs to be manually
updated if a new JS file is added or existing file removed.